Update for group name to path GitLab API change.
As of GitLab version 10.8.3, the API now expects the full path to the
namespace, not just the name.

See https://gitlab.rc.uab.edu/jelaiw/ccts-bmi-incubator/issues/68.

Update variable names and code comments for clarity.

Remove old code comments (where the previous behavior of the API was
inconsistent with the API documentation) that explained what we had
observed.